+++ src/main/java/de/stamme/basicquests/ServerInfo.java
@@ -70,7 +70,7 @@ public class ServerInfo implements Serializable {
     private void cleanMap(HashMap<QuestData, LocalDateTime> map) {
         for (Map.Entry<QuestData, LocalDateTime> entry: map.entrySet()) {
             long secondsAgo = Duration.between(entry.getValue(), LocalDateTime.now()).getSeconds();
-            if (secondsAgo > 86400)
+            if (secondsAgo > 604800) // 1 Week
                 map.remove(entry.getKey());
         }
     }