+++ src/main/java/de/stamme/basicquests/commands/HideQuestsCommand.java
@@ -1,5 +1,7 @@
 package de.stamme.basicquests.commands;
 
+import de.stamme.basicquests.Config;
+import net.md_5.bungee.api.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
@@ -15,6 +17,11 @@ public class HideQuestsCommand implements CommandExecutor {
 	@Override
 	public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String label, String[] args) {
 		if (!(sender instanceof  Player)) return false;
+
+		if (Config.isScoreboardDisabled()) {
+			sender.sendMessage(ChatColor.RED + "This feature is disabled");
+			return true;
+		}
 
 		QuestPlayer questPlayer = Main.getPlugin().getQuestPlayer((Player) sender);
 		if (questPlayer == null) return true;